Refactor: Pull out `serialize_device` to Device class
- Ensures relevant schema serialisation is with associated definition.
- Note: type definition on serialize must be in quotes as it is a 'forward reference'
