+++ tests/consumer_tests/routers/test_inventory.py
@@ -1,6 +1,7 @@
 import pytest
 
 from consumer.routers import inventory as router_inventory
+from consumer.schemas.device import Device
 from consumer.services import inventory
 
 
@@ -70,7 +71,7 @@ def test_device_history_with_device_in_use_success(
     response_row, device_history, client, monkeypatch
 ) -> None:
     async def mock_device_by_id(device_id: str):
-        return router_inventory.serialize_device(response_row)
+        return Device.serialize(response_row)
 
     monkeypatch.setattr(router_inventory, "device_by_id", mock_device_by_id)
 
@@ -89,7 +90,7 @@ def test_device_history_with_device_in_use_fail(
     response_row, device_history, client, monkeypatch
 ) -> None:
     async def mock_device_by_id(device_id: str):
-        return router_inventory.serialize_device(response_row)
+        return Device.serialize(response_row)
 
     monkeypatch.setattr(router_inventory, "device_by_id", mock_device_by_id)
 
@@ -113,7 +114,7 @@ def test_serialize_device_required_params(key, response_row) -> None:
     del response_row[key]
 
     with pytest.raises(KeyError):
-        router_inventory.serialize_device(response_row)
+        Device.serialize(response_row)
 
 
 serial_optional_params = [
@@ -127,6 +128,6 @@ serial_optional_params = [
 def test_device_serial_optional_params(key, child, response_row) -> None:
     del response_row[key][child]
 
-    result = router_inventory.serialize_device(response_row)
+    result = Device.serialize(response_row)
 
     assert not result.dict()[key]
